pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "lcr"
version = "0.0.1"
description = "Long Context Retrieval ITU Bachelor Project"
authors = [
  { name = "Maciej Jalocha, William Michelsen" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "beautifulsoup4>=4.14.3",
    "datasets>=4.8.4",
    "loguru",
    "openai>=2.30.0",
    "pip",
    "python-dotenv",
    "pytrec-eval>=0.5",
    "ruff",
    "tqdm",
    "typer",
]
requires-python = "~=3.11.0"
[project.optional-dependencies]
cpu-deps = [ # to be run with --extra-index-url https://download.pytorch.org/whl/cpu
    "sentence-transformers",
    "torch",
]


[tool.ruff]
line-length = 99
src = ["lcr"]
include = ["pyproject.toml", "lcr/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["lcr"]
force-sort-within-sections = true

